requirements.txt
alembic==1.15.2
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1733332318622/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
bleach==6.2.0
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1725267563793/work
catboost==1.2.8
category_encoders==2.8.1
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1746569525376/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560577695/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1735929714516/work
cloudpickle==3.1.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
colorlog==6.9.0
colour==0.1.5
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
contourpy @ file:///Users/runner/miniforge3/conda-bld/contourpy_1744743111830/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1744321336370/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
dtreeviz==2.2.2
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1733569351617/work
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1743732266240/work
future @ file:///home/conda/feedstock_root/build_artifacts/future_1738926421307/work
graphviz==0.20.3
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
h2o==3.46.0.7
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
imbalanced-learn @ file:///home/conda/feedstock_root/build_artifacts/imbalanced-learn_1734717084342/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
iniconfig==2.1.0
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1719845458456/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1744033187/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1733736026804/work
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
kaggle==1.7.4.2
kagglehub @ file:///home/conda/feedstock_root/build_artifacts/kagglehub_1739488418739/work
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1725459292489/work
lightgbm==4.6.0
llvmlite==0.44.0
Mako==1.3.10
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mljar-scikit-plot==0.3.12
mljar-supervised==1.1.17
munkres==1.1.4
narwhals==1.37.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
numba==0.61.2
numpy==1.26.4
opencv-python==4.11.0
opencv-python-headless==4.11.0
optuna==4.3.0
optuna-integration==4.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas @ file:///Users/runner/miniforge3/conda-bld/pandas_1744430513098/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
patsy @ file:///home/conda/feedstock_root/build_artifacts/patsy_1733792384640/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1735929746402/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1742485085/work
plotly==6.0.1
pluggy==1.5.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1737453357274/work
protobuf==6.30.2
psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1740663164037/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1743089729650/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
pytest==8.3.5
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-slugify==8.0.4
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML==6.0.2
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1743831272844/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
scikit-learn @ file:///Users/runner/miniforge3/conda-bld/scikit-learn_1736496791658/work/dist/scikit_learn-1.6.1-cp313-cp313-macosx_11_0_arm64.whl#sha256=5b69dde9fbe5319050a52157ae14972bce36d5255fc98ae69812194282a415af
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy-split_1739790783802/work/dist/scipy-1.15.2-cp313-cp313-macosx_11_0_arm64.whl#sha256=b07c2380f54734e2bc5d19420a293aabf6bcd5a34beba80eace1959b32b5edde
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1733730015268/work
setuptools==78.1.0
shap==0.47.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sklearn-compat @ file:///home/conda/feedstock_root/build_artifacts/sklearn-compat_1734794177910/work
slicer==0.0.8
SQLAlchemy==2.0.40
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
statsmodels @ file:///Users/runner/miniforge3/conda-bld/statsmodels_1727986717451/work
tabulate @ file:///home/conda/feedstock_root/build_artifacts/tabulate_1733589744265/work
text-unidecode==1.3
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1732615933794/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1744302253/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1744323578849/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
webencodings==0.5.1
wordcloud==1.9.4
xgboost==3.0.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
zstandard==0.23.0
